Overall performance improvements and memory saving refactorings. The UDPPacket has been removed, as it doesn't seem to be important anymore. The rest of the code has refactorings as well.
Bugfix regarding N-*.pcap entries for a Drone-command containing N-commands, instead of one *.pcap entry.

The beginning of a cmd-packet is no guess anymore, but will be calculated based on the provided wifi-header.

Old lava-code removed. This commit contains a "play" button which will crash the application. Ignore it - dont press it. Furthermore, the Drone packet listener isn't working in this commit as well, since that is currently in the refactoring process. Dont use this commit, it is more or less a savepoint for me.